--- src/main/java/com/splitface/tattoo/Repositories/ArtistRepository.java
@@ -1,7 +1,0 @@
-package com.splitface.tattoo.Repositories;
-
-import org.springframework.data.repository.CrudRepository;
-
-public interface ArtistRepository extends CrudRepository<Artist,Long>{
-
-}
