--- src/main/java/com/splitface/tattoo/Service/ArtistService.java
@@ -1,6 +1,0 @@
-package com.splitface.tattoo.Service;
-
-public interface ArtistService {
-
-    List<Artist> getAllArtist();
-}
