+++ src/main/java/com/splitface/tattoo/service/ArtistService.java
@@ -2,7 +2,11 @@ package com.splitface.tattoo.service;
 
 import com.splitface.tattoo.models.Artist;
 
+import java.util.List;
+
 public interface ArtistService {
 
     Artist createArtistInDb(Artist artist);
+
+    List<Artist> getAllArtist();
 }
