+++ src/main/java/com/splitface/tattoo/service/serviceImpl/ArtistServiceImpl.java
@@ -6,6 +6,9 @@ import com.splitface.tattoo.service.ArtistService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.List;
+
 @Service
 public class ArtistServiceImpl implements ArtistService {
 
@@ -19,5 +22,11 @@ public class ArtistServiceImpl implements ArtistService {
     @Override
     public Artist createArtistInDb(Artist artist) {
         return null;
+    }
+
+    public List<Artist> getAllArtist() {
+        List<Artist> artistList = new ArrayList<>();
+        artistRepository.findAll().forEach(artistList::add);
+        return artistList;
     }
 }
